MDL-51569 mod_choice: Check choice availability prior to do any action

+++ mod/choice/view.php
@@ -34,7 +34,10 @@
 
     $context = context_module::instance($cm->id);
 
-    if ($action == 'delchoice' and confirm_sesskey() and is_enrolled($context, NULL, 'mod/choice:choose') and $choice->allowupdate) {
+    list($choiceavailable, $warnings) = choice_get_availability_status($choice);
+
+    if ($action == 'delchoice' and confirm_sesskey() and is_enrolled($context, NULL, 'mod/choice:choose') and $choice->allowupdate
+            and $choiceavailable) {
         if ($answer = $DB->get_record('choice_answers', array('choiceid' => $choice->id, 'userid' => $USER->id))) {
             $DB->delete_records('choice_answers', array('id' => $answer->id));
 
@@ -63,6 +66,12 @@
                 redirect("view.php?id=$cm->id");
             }
         }
+
+        if (!$choiceavailable) {
+            $reason = current(array_keys($warnings));
+            throw new moodle_exception($reason, 'choice', '', $warnings[$reason]);
+        }
+
         // Redirection after all POSTs breaks block editing, we need to be more specific!
         $answer = optional_param('answer', '', PARAM_INT);
         if ($answer) {
